requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
drf-yasg==1.21.11
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
whitenoise==6.11.0
